Refactor Slack Notifications into Lib
