+++ src/mavedb/server_main.py
@@ -1,8 +1,4 @@
-import json
 import logging
-import os
-import sys
-import traceback
 
 import uvicorn
 from fastapi import FastAPI
@@ -11,7 +7,6 @@ from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.openapi.utils import get_openapi
 from requests import Request
-from slack_sdk.webhook import WebhookClient
 from sqlalchemy.orm import configure_mappers
 from starlette import status
 from starlette.responses import JSONResponse
@@ -40,6 +35,7 @@ from mavedb.routers import (
 )
 from mavedb.lib.exceptions import AmbiguousIdentifierError, NonexistentIdentifierError, MixedTargetError
 from mavedb.lib.permissions import PermissionException
+from mavedb.lib.slack import send_slack_message
 
 logging.basicConfig()
 # Un-comment this line to log all database queries:
@@ -135,45 +131,11 @@ def customize_validation_error(error):
     return error
 
 
-def find_traceback_locations():
-    _, _, tb = sys.exc_info()
-    return [
-        (fs.filename, fs.lineno, fs.name)
-        for fs in traceback.extract_tb(tb)
-        # attempt to show only *our* code, not the many layers of library code
-        if "/mavedb/" in fs.filename and "/.direnv/" not in fs.filename
-    ]
-
-
 @app.exception_handler(Exception)
 async def exception_handler(request, err):
     logger.error("Uncaught exception", exc_info=err)
+    send_slack_message(err=err, request=request)
 
-    text = json.dumps({
-        "type": err.__class__.__name__,
-        "exception": str(err),
-        "location": find_traceback_locations(),
-        "client": str(request.client.host),
-        "request": f"{request.method} {request.url}",
-    })
-
-    slack_webhook_url = os.getenv("SLACK_WEBHOOK_URL")
-    if slack_webhook_url is not None and len(slack_webhook_url) > 0:
-        client = WebhookClient(url=slack_webhook_url)
-        response = client.send(
-            text=text,
-            blocks=[
-                {
-                    "type": "section",
-                    "text": {
-                        "type": "plain_text",
-                        "text": text
-                    },
-                }
-            ],
-        )
-    else:
-        print(f"EXCEPTION_HANDLER: {text}")
     return JSONResponse(status_code=500, content={"message": "Internal server error"})
